refactor(frontend/src/features/commands
/api/apiCommands.js):

move apiCommands.js to features/commands/api for

better feature organization
